vision_quote_pipeline/requirements.txt
numpy>=1.24
pandas>=2.0
Pillow>=10.0
opencv-python>=4.9
pyyaml>=6.0
tqdm>=4.66
faiss-cpu>=1.8.0
open_clip_torch>=2.24.0
torch>=2.1
click>=8.1
python-dotenv>=1.0
pyarrow>=15.0
